【RichEditor】修复告警问题

+++ frameworks/bridge/declarative_frontend/jsview/js_richeditor.cpp
@@ -2556,7 +2556,7 @@ void JSRichEditorBaseController::ParseTextDirection(const JSRef<JSObject>& style
     }
 
     int32_t index = textDirectionObj->ToNumber<int32_t>();
-    auto isNormalValue = index >= 0 && index < TEXT_DIRECTIONS.size();
+    auto isNormalValue = index >= 0 && index < static_cast<int32_t>(TEXT_DIRECTIONS.size());
     if (!isNormalValue) {
         style.textDirection = TEXT_DIRECTIONS[INHERIT_INDEX];
         return;
